prepping for 2.0.0 release
