--- Tools/GitVersion/GitVersion.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0"?>
-<doc>
-    <assembly>
-        <name>GitVersion</name>
-    </assembly>
-    <members>
-    </members>
-</doc>
