--- Tools/GitVersion/LibGit2Sharp.dll.config
@@ -1,4 +1,0 @@
-﻿<configuration>
-    <dllmap os="linux" cpu="x86-64" wordsize="64" dll="git2-572e4d8" target="lib/linux-x64/libgit2-572e4d8.so" />
-    <dllmap os="osx" cpu="x86,x86-64" dll="git2-572e4d8" target="lib/osx/libgit2-572e4d8.dylib" />
-</configuration>
