+++ src/Sample/Tests/Tests.csproj
@@ -18,9 +18,9 @@
 	</ItemGroup>
 
 	<ItemGroup>
-		<PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="7.0.9" />
-		<PackageReference Include="Microsoft.AspNetCore.TestHost" Version="7.0.9" />
-		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.6.3" />
+		<PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="7.0.10" />
+		<PackageReference Include="Microsoft.AspNetCore.TestHost" Version="7.0.10" />
+		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.7.1" />
 		<PackageReference Include="Shouldly" Version="4.2.1" />
 		<PackageReference Include="SnowStorm.Extensions" Version="1.0.0" />
 		<PackageReference Include="xunit" Version="2.5.0" />
