+++ src/SnowStorm/Queries/IQueryableProvider.cs
@@ -1,0 +1,9 @@
+﻿using System.Linq;
+
+namespace SnowStorm.QueryExecutors
+{
+    public interface IQueryableProvider
+    {
+        IQueryable<T> Query<T>() where T : class;
+    }
+}
