--- src/SnowStorm/QueryExecutors/IQueryableProvider.cs
@@ -1,9 +1,0 @@
-﻿using System.Linq;
-
-namespace SnowStorm.QueryExecutors
-{
-    public interface IQueryableProvider
-    {
-        IQueryable<T> Query<T>() where T : class;
-    }
-}
